+++ central/src/menu.c
@@ -115,10 +115,10 @@ void print_data(Output output)
     wrefresh(menu_win);
 }
 
-void print_temperature(float temperature, float humidity) {
-    printf("TEMPERATURAAAAAA: %f\n", temperature);
-    mvwprintw(menu_win, 10, 2, "Temperatura: %f", temperature);
-    mvwprintw(menu_win, 11, 2, "Umidade: %f", humidity);    
+void print_temperature(double temperature, double humidity) {
+    printf("TEMPERATURAAAAAA: %lf\n", temperature);
+    mvwprintw(menu_win, 10, 2, "Temperatura: %lf", temperature);
+    mvwprintw(menu_win, 11, 2, "Umidade: %lf", humidity);    
     wrefresh(menu_win);
 }
